+++ pages/_app.tsx
@@ -9,8 +9,7 @@ import Router from 'next/router'
 import NProgress from 'nprogress'
 import { I18nProvider } from '@lingui/react'
 import { i18n } from '@lingui/core'
-import { activate, activateAndSetCookie } from '../utils/i18n'
-import { detect, fromUrl, fromCookie } from '@lingui/detect-locale'
+import { activate, detectAndSetLocale } from '../utils/i18n'
 
 init()
 
@@ -36,13 +35,7 @@ function MyApp({
     }
     // Activate translation on client side after locale detection
     // TODO: To be implemented
-    const DEFAULT_FALLBACK = () => 'en'
-    const locale = detect(
-      fromUrl('lang'),
-      fromCookie('current_locale'),
-      DEFAULT_FALLBACK
-    )
-    activateAndSetCookie(locale)
+    detectAndSetLocale()
   }, [])
 
   return (
